requirements.txt
discord.py
pickle

